Use less memory for TZ string timestamps
32 bits aren't enough to hold the start times of 400 years, but
they're ample to hold start times relative to the beginning of the
year.  Let's cut our memory usage roughly in half by recording the
start time relative to the start of the year and keeping one common
table of the start timestamps of each year.

+++ lib/tz64file.h
@@ -52,7 +52,7 @@ struct tz64 {
     const int64_t *rev_leap_ts;
     const int32_t *leap_secs;
     const char *desig;
-    const int64_t *extra_ts;
+    const int32_t *extra_ts;
 };
